\name{murnau.X}
\docType{data}
\alias{murnau.X}
\alias{murnau.Y}
\alias{murnau.W}
\title{Hyperspectral reflectance and plant attributes}
\description{
This data gives reflectance from raised bogs and fens (\code{murnau.X}),
the corresponding wavelengths (\code{murnau.W}) and associated vegetation 
attributes (\code{murnau.Y}: plant adaptation to stress). More detail is given
by Schmidtlein et al. (2012).
}
\usage{data(murnau.X)}
\format{Matrix containing 40 observations and 105 spectral bands 
(\code{murnau.X}), a vector with 105 wavelengths corresponding to bands 
(\code{murnau.W}) and a vector with 40 observations of plant attributes 
(\code{murnau.Y}).}
\source{
  Schmidtlein, S., Feilhauer, H., Bruelheide, H. (2012): Mapping plant
  strategy types using remote sensing. \emph{Journal of Vegetation Science}
  \bold{early view}.
}
\keyword{datasets}